+++ src/View/Tag.php
@@ -2,6 +2,7 @@
 
 namespace Expressionengine\Coilpack\View;
 
+use Illuminate\Support\Collection;
 use Illuminate\Support\Str;
 
 abstract class Tag
@@ -14,11 +15,42 @@ abstract class Tag
     protected $arguments = [];
 
     /**
+     * A collection of parameters available on the tag
+     *
+     * @var Collection|null
+     */
+    protected $parameters;
+
+    /**
      * Run the tag logic to produce the output
      *
      * @return mixed
      */
     abstract public function run();
+
+    /**
+     * Get a collection of defined Parameters
+     *
+     * @return Collection
+     */
+    public function parameters(): Collection
+    {
+        if (! $this->parameters) {
+            $this->parameters = collect($this->defineParameters())->keyBy('name');
+        }
+
+        return $this->parameters;
+    }
+
+    /**
+     * Define the parameters available on this Tag
+     *
+     * @return array
+     */
+    public function defineParameters(): array
+    {
+        return [];
+    }
 
     /**
      * Set the arguments to be used by the tag
@@ -62,14 +94,34 @@ abstract class Tag
         return $this->{'set'.Str::studly($key).'Argument'}($value);
     }
 
-    public function hasArgument($key)
+    /**
+     * Determine whether or not an argument was set
+     *
+     * @param  string  $key
+     * @return bool
+     */
+    public function hasArgument(string $key)
     {
         return isset($this->arguments[$key]);
     }
 
-    public function getArgument($key)
+    /**
+     * Retrieve an argument value
+     * If a value is not set the parameter's default value will be returned
+     *
+     * @param  string  $key
+     * @return mixed
+     */
+    public function getArgument(string $key): mixed
     {
-        return $this->arguments[$key];
+        if ($this->hasArgument($key)) {
+            return $this->arguments[$key];
+        }
+
+        // If a default value was defined on the parameter use it
+        if ($this->parameters()->has($key)) {
+            return $this->parameters()->get($key)->defaultValue ?? null;
+        }
     }
 
     /**